+++ .env.dist
@@ -1,2 +1,4 @@
-JSON_DB_PATH = .data/equipos.json
-PORT = 8080
+JSON_DB_PATH = secret_path
+SESSION_SECRET = secret_session
+PORT = secret_port
+UPLOAD_FILE_PATH = secret_path_upload
