+++ data/equipos.json
@@ -390,13 +390,13 @@
     "lastUpdated": "Wed Jul 14 2021 14:07:04 GMT-0300 (hora estándar de Argentina)"
   },
   {
-    "id": "709f64a2-ce14-430f-b5ca-456ef7716f87",
+    "id": "9b270e71-5ff9-48b8-be7f-c99cde9e7f55",
     "area": {
       "name": "Argentina"
     },
     "name": "River",
     "crestUrl": "/images/CA_river_plate_logo.svg.png",
     "address": "Av. Pres. Figueroa Alcorta 7597, C1428 CABA",
-    "lastUpdated": "Fri Jul 23 2021 13:55:32 GMT-0300 (hora estándar de Argentina)"
+    "lastUpdated": "Fri Jul 23 2021 16:25:24 GMT-0300 (hora estándar de Argentina)"
   }
 ]
