+++ modulos/club/configDI/container-di.js
@@ -6,10 +6,41 @@ const {
 const fs = require('fs');
 const { v4: uuidv4 } = require('uuid');
 const session = require('express-session');
-
-const jsonDataBasePath = process.env.JSON_DB_PATH;
+const multer = require('multer');
 
 const { ClubController, ClubService, ClubRepository } = require('../module');
+
+function configureJSONDatabase() {
+  return process.env.JSON_DB_PATH;
+}
+
+function configureUuidv4() {
+  return uuidv4;
+}
+
+function configureMulter() {
+  const storage = multer.diskStorage({
+    destination(req, file, cb) {
+      cb(null, process.env.UPLOAD_FILE_PATH);
+    },
+    filename(req, file, cb) {
+      cb(null, file.originalname);
+    },
+  });
+  const uploadConfigure = {
+    storage,
+    limits: { fileSize: 1000000 },
+    fileFilter(req, file, cb) {
+      if (file.mimetype === 'image/png' || file.mimetype === 'image/jpeg' || file.mimetype === 'image/jpg') {
+        return cb(null, true);
+      }
+      return cb('Error: Solo se soportan imagenes!');
+    },
+  };
+
+  const upload = multer(uploadConfigure);
+  return upload;
+}
 
 function configureSession() {
   const oneWeekInSeconds = 604800000;
@@ -22,12 +53,14 @@ function configureSession() {
   };
   return session(sessionOptions);
 }
+
 function addCommonDefinitions(container) {
   container.addDefinitions({
     fs,
-    uuidv4,
+    uuidv4: factory(configureUuidv4),
+    Multer: factory(configureMulter),
     Sessions: factory(configureSession),
-    JSONDatabase: jsonDataBasePath,
+    JSONDatabase: factory(configureJSONDatabase),
 
   });
 }
