+++ routes/crud-routes.js
@@ -1,32 +1,15 @@
 /* eslint-disable eqeqeq */
-const multer = require('multer');
 const express = require('express');
 const container = require('../modulos/club/configDI/container-di')();
 /**
  * @type {import('../modulos/club/controllers/clubController')} clubController
  */
 const clubController = container.get('ClubController');
+const upload = container.get('Multer');
 const router = express.Router();
+
 router.use(container.get('Sessions'));
-// Storage Settings
-const storage = multer.diskStorage({
-  destination(req, file, cb) {
-    cb(null, './uploads/images');
-  },
-  filename(req, file, cb) {
-    cb(null, file.originalname);
-  },
-});
-const upload = multer({
-  storage,
-  limits: { fileSize: 1000000 },
-  fileFilter(req, file, cb) {
-    if (file.mimetype === 'image/png' || file.mimetype === 'image/jpeg' || file.mimetype === 'image/jpg') {
-      return cb(null, true);
-    }
-    return cb('Error: Solo se soportan imagenes!');
-  },
-});
+
 // Home
 router.get('/', clubController.clubIndex.bind(clubController));
 // Agregar
